requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
filelock==3.18.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
platformdirs==4.3.7
python-dotenv==1.1.0
redis==6.0.0
requests==2.32.3
spotipy==2.25.1
SQLAlchemy==2.0.40
typing_extensions==4.13.1
urllib3==2.3.0
virtualenv==20.30.0
Werkzeug==3.1.3
